+++ templates/__common__/app/scripts/utils/d3-base.js
@@ -1,6 +1,5 @@
 import { axisBottom, axisLeft } from 'd3-axis';
 import { scaleLinear } from 'd3-scale';
-import { select } from 'd3-selection';
 
 export default function createBase({
   margin,