+++ templates/graphic/app/scripts/embeds/frames.js
@@ -1,19 +1,18 @@
 // packages
-import { initFrame } from '@newswire/frames';
+import { initFrame, sendFrameHeight } from '@newswire/frames';
 
 // local
 import debounce from '../utils/debounce';
 
 /**
- * Sets up an instance of a frame.
+ * Sets up an instance of a frame and executes a function, i.e. a function to render the graphic.
  *
- * @param {Function} [fn] An optional function to be called on resize
+ * @param {Function} [fn] An optional function to be called on load and resize
  */
 export function frameLoader(fn) {
-  // activate the frame
+  // initialize the frame no matter what
   initFrame();
 
-  // only needed if a function was provided
   if (fn) {
     // set up a debounced version
     const debouncedFn = debounce(fn, 300);
@@ -21,7 +20,10 @@ export function frameLoader(fn) {
     // resize listener
     window.addEventListener('resize', debouncedFn);
 
-    // initial kickoff
-    debouncedFn();
+    // if fn() renders a graphic, it needs to come before the frame height is sent
+    new Promise(resolve => {
+      fn();
+      resolve(sendFrameHeight());
+    });
   }
 }