+++ CHANGELOG.md
@@ -10,6 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 
+## [2.0.0] - 2019-09-25
+### Added
+- `app/styles/raw-plugin-styles.html` - add styles snippet that goes into the CSS content section of Raw Plugins
+- `app/static.html` - template for non-scripted graphics
+- `app/scripts/static.js` - pack for non-scripted graphics
+
+### Changed
+- Switched from using Pym to `frames` library #17
+
+### Removed
+- `app/scripts/embeds/pym.js`
+
 ## [1.2.0] - 2019-09-19
 ### Fixed
 - Fixed security vulnerability introduced by `lodash` #10